Refactor auth models
